Reorganize the Fem3D example
To make the Fem 3D example more instructive, the Cube and Tetrahedrons
are built up seperatly step by step. This introduces some code
duplication, but I think it is instructive to see each step clearly.

In reality, users are going to use either cubes or tetrahedrons, not
both. So to see one function that does the complete setup, I think, is
useful.