Remove gmock/gtest from ThirdParty
Remove Google Mock and Google Test from third party and treat them like
a regular dependancy. However, since we need to actually build
gmock/gtest with our project, according to Google's recommendation, we
do things slightly different. The FindGoogleMock looks for the gmock
source directory (debian default location and environment variable).
Then, OpenSurgSim adds gmock with an add_subdirectory, so it is adding
like any other source directory, but it is external to our source tree,
which is fine. Now gmock is built with OpenSurgSim, as before, but it
lives outside of OpenSurgSim, and is installed like any other
dependancy.