Fix Octree-Capsule Test on Win7
The octree contact calculation tests were failing on Win7 in release
mode. This, I think, is due to not using the Eigen::DontAlign flag on
two array's that were being used in intersectionsSegmentBox in Geometry.h

This change, just uses the same Matrix Options as the vector's passed
into the function. So if they are Eigen::DontAlign (as is the default in
our Vectors and Matrices) the two arrays will be DontAlign.

refs #12