Create BUILD_DEVICE_* for each device
Instead of building devices when the appropriate SDK is found (ie Novint
hdal, OpenHaptics, etc), a BUILD option is added for each device
(BUILD_DEVICE_NOVINT, BUILD_DEVICE_PHANTOM, etc). Now if you enable the
device, it will check fot he appropriate SDK and give you an error if it
isn't there.

Also, removed the option to disable the running of unit tests. This was
used to build tests for devices, but not run them. In the case where the
SDK was installed but the device was connected. Instead we will use the
option in CTest (CTEST_CUSTOM_TESTS_IGNORE) to disable the running of
tests.