Moved tests of makeRigidTransform  out of MatrixTests
The makeRigidTransform for making a view matrix was incorrectly tested
in MatrixTests. It should be in RigidTransformTests. As a side effect
these tests were never being run, since MatrixTests test Eigen, and
these tests are off by default. I moved it into RigidTrasnsformTests.
The test was also incorrectly calling makeViewMatrix, when it should be
makeRigidTransform.

However, the test does fail, and that needs to be addressed.