Fix generating export targets file
When CTest and CDash support was added (commit
ee33167ab0b381cccefad21b681cad61b9939244) creating of the export targets
file in the build tree (OpenSurgSimTargets.cmake) was removed. In
addition, the project was no longer registered with CMAke.  This change
adds it back for using OpenSurgSim from the build-tree.