Adds output component to VirtualToolCoupler.
The output component is sent the nominal force, torque, stiffness,
damping, and the current pose of the input.  This information is
sufficient for the output device to calculate forces and torques in its
update loop.

The VTC now relies on the input device to provide a velocity,
otherwise it assumes no input velocity.

Updates the constructor signature to be single-argument.