Zeroes the Novint forces/torques if a fatal error occurs beforehand.
If the HDAL reports a fatal error when the Scaffold gets the grip
attributes, or when gravity compensation is set, this will zero
the forces and torques.  Previously the code would output the
forces/torques calculated in the previous update, which could cause
uncontrolled motions.

The torque data is passed around as a Vector4d in the DeviceData.
torqueBuffer/torqueValue hold the torque command counts per motor axis.
The fourth value is for the jaw torque.