Adds functions to set the velocities for a RigidRepresentation.
Since the velocities are used in free motion to update the pose of a
RigidRepresentation; with these functions a Behavior can exactly control
the next pose.  Note that setPose is non-functional for
RigidRepresentations, and making it functional would be incorrect
because then any residual velocity would cause the representation to
move away from the pose in the free motion step.

Currently this is being used in a side project to move the
representation of a tool to the pose being received from an
InputComponent.