Fix mesh inertia matrix calculation
There was a bug in the previous code to compute the 1st entry of the inertia matrix:
we had in the code:
secondMoment(0, 0) -= volume * (r[1]*r[1] + r[0]*r[0]);
while this part of Ixx only depends on y and z axis terms => the fix was
secondMoment(0, 0) -= volume * (r[1]*r[1] + r[2]*r[2]);

I found a paper from David Eberly improving the Mirtich algorithm in terms of both speed and clarity, so i switched them.
And also added checks for retrieving the volume and second moment of volume to test their validity.