Make Spring/MassSpringRep API similar to FemElement/FemRep API
Spring is not holding the force vector nor the stiffness/damping matrices anymore, it build them and add their respective contributions directly into a bigger system, like FemElement is doing.
This simplifies the code a little, especially when assembling the system, it is done more naturally on the MassSpringRepresentation layer, it standardize the API, avoid storing unnecessary data in Spring.
The MassSpringRep code is now very close to FemRep code, the addRayleighDampingForce method is now using the same construct as in FemRep, it takes 3 boolean to improve performances when the proper matrices (D or M,K) have been build.
+
rename variable with full name (rayleighMass and rayleighStiffness) in MassSpringRep