Add Linear Version of the Ode Solvers
Create a ode solvers that assume the system it is solving is linear.
During initialization it uses the non-linear version of the solver, then
uses cached values from then on.

Modified the unit tests to test both the linear and non-linear solvers
side by side.

Changed ExampleStapling to use the Linear Implicit solver, increases
speed greatly.