Calculate sensible VTC parameters automatically
This change calculates sensible vtc parameters if non are provided. The
damping and stiffness values (linear and angular) are put into
OptionalValues. During wake up, they are calculated if non are provided
based on the rigid representation properties (mass and moment of
inertia). This method provides similar response of the vtc regardless of
the shape and mass of the controled object.

Also, removed setting these values in our examples, and use the
automatic calculation.