Improve logging during constraint generation
Distinguish between when a collision representation doesn't have a
physics representation (which is fine) from an error making a constraint
(which is probably not fine).

Also, add a log message when components are added to managers. The
graphics manager was already doing this. Speaking of the graphics
manager, it was creating a logger when it didn't need to. It's super
class is doing it.