VTC's default parameters are critically damped
Now, the vtc's default parameters are set such that the vtc is
critically damped. If neither of the parameters are given (stiffness and
damping) they are both calculated to provide good critically damped
response. If one of the parameters is given, the other is calculated
such that the system is critically damped. Nothing is calculated if both
parameters are set. This is done the same way with the angular
component of the vtc.