Add test of vtc default parameters, cleanup
Added unit tests of the default vtc parameters. The unit tests verify
that the default parameters provide a critically damped system by
checking the settling time and exponential behavior.

Added comments about how the default parameters are calculated, also
added comments to the setters of the parameters indicating they will
automatically be calculated if not provided.

Finally removed comments on inherited functions that already have
comments in the parent class.