Representation's pose dependant on the SceneElement
Calling getPose on a Representation now returns the combined pose of the
Representation (get/setLocalPose) and the SceneElement's pose. The local
pose is the pose of the representation with respect to the SceneElement.

Also, now physics representations can 'drive' the scene element. Using a
DriveElementBehavior the pose of the physic representation will provide
the pose of the element. So other representations' poses will be updated
as the physics representation updates the element pose. There is
similarly now a DriveElementFromInputBehariour that allows an input
component to drive an element. These two behaviors remove the need for
the TransformPoseBehaviors.

Removed MockObjects from Blocks since the only place these MockObjects
were being used was actually in the Framework Unitests. So moved the
objects from blocks MockObjects in the MockObjects in Framework.

Updated all unit tests and examples to reflect these changes.