Add Camera to View Element
Add a camera to the ViewElement so it can be posed more intuitively.
Since Representation (which a Camera is) can be posed by there elements,
this provides a SceneElement for Camera's to live in.  This should be
easier than the default camera (that does not live in a SceneElement).
Since the ViewElement (which is a SceneElement) has a pose, you can set
the ViewElement's pose, and this will move the camera. Instead of using
the less intuitive setLocalPose of the camera, and assuming that it
isn't accidently in another SceneElement that could get posed.