Remove DriveElementBehavior
Instead of using a behavior that drives the SceneElement with a Physics
Representations instead just have the Physiscs Representations drive the
SceneElement directly. Also, added a flag to PhysicsRepresentation so
you can turn off this ability. But the default behavior, is that if a
PhysiscRepresentation is added to a SceneElement in will be changeing
the SceneElement pose in afterUpdate.