Fix passing aligned Eigen types by value
With aligned Eigen types, you cannot pass by value. This commit changes
the couple of cases where we were passing by value to const reference.

Unfortunatly, google tests EXPECT_EQ causes a pass by value, and we
cannot change that. So, switched it to EXPECT_TRUE in the couple of
cases that were causing alignment errors. EXPECT_EQ wasn't printing
useful information in the couple of cases it was used with Eigen types,
so we aren't losing any useful information by making the switch.