Remove boundary condition (BC) compliance to avoid LCP violating BC.
The boundary conditions are set by zeroing out all rows and columns of the fixed dofIds.
We also set the diagonal element of these dofIds to 1, to make sure the resulting matrix is invertible.
By doing so, the compliance matrix (inversed matrix) has the same structure, 0 all over these lines and columns except a 1 on the diagonal of these dofIds.

This 1 causes trouble when solving a contact involving a node fixed by boundary conditions.
Imagine a contact on a fixed node with a tool, then after free motion, we solve the contact by solving the following LCP:
H = dt.(0...0 nx ny nz 0...0) for the given dofID of the contacting node (a fixed node with boundary conditions)
H.C.H^t = dt.(0...0 nx ny nz 0...0).(Id on the dofId we are looking at).dt.(0...0 nx ny nz 0...0)^t
= dt.(0...0 nx ny nz 0...0). dt.(0..0 nx ny nz 0..0)^t
=dt^2.(nx^2 + ny^2 + nz^2) = dt^2
Which means that the LCP will solve dt^2.lambda + initialViolation >= 0 and will find a non null force for this problem that will push on the fixed node.

To fix that problem, we can:
1) applyBoundaryConditionsToVector(displacement) in the applyDofCorrection..
OR
2) we can remove all compliances in the m_compliance matrix (zero for diagonal element).

The later is the proper choice because a contact will involve 2 entities, so the LCP system will get a compliance from both objects,
if the fixed dof cannot move because of a boundary condition, the LCP should not consider this node to have a compliance, otherwise it will screw up the correction on the other object.
Indeed, the other object will think that our fixed node can actually move and will compute improper displacement based on this wrong assumption.