Use mass center in Rigid's free motion
The mass center was not being used to calculate rigid body dynamics.
Torques were in effect being applied around (0,0,0), which for most of
the shapes is the mass center. But for meshes, the mass center is likely
to be somewhere else. This change shifts the frame of reference to the
center of mass.