Fix the RigidCollisionRep::update using the correct pose
Before, if you setup a scene with 1 fixed plane, 2 rigid sphere falling (1 SphereShape and 1 MeshShape),
the shape will fall and stay in contact while the mesh wil fall and bounce.
Now the have the exact same behavior, they both stays in contact.

Same thing with a staple dropped on a plane, it stays in contact.