Increase timing for RigidRep sliding on planes to 13s (from 10s)
This is  to exhibit some behavior of the boxes on the planes (both mesh and shape).
They get more and more torque with each collision on the other plane the are sliding of, which makes them rotate on themselves.
It is what i would expect, but good to show that happening through that render test.