Bug: Missing include in TetrahedronMesh.h
This does not show up anywhere because we only have the unit test using that so far.
And it includes all the necessary header files prior to including TetrahedronMesh.h, which solves the issue.
But in a different project, if someone includes TetrahedronMesh.h, it gets errors for missing MeshElement, Vertices, std::vector...