The VTC takes an attachment point where the input connects to the tool. Often the input position provided by a haptic device should correspond to the tip of the tool, but physics does all torques about the mass center, so we need to convert somewhere.
The forces and torques experienced about the mass center of the physical
tool should be equal to those set in addExternalForce and
addExternalTorque for application to the mass center of the simulated
tool.

The device scaffold could do the conversion by adjusting the kinematic
calculations, but then each time the mesh representing the tool changes
we would need to update the device.  Similarly, a device filter could do
the calculations but then it needs to know about the mesh shape.
The VTC already knows about all the pieces and can do the calculations.

The stapler in StaplingExample should now rotate about the tip as the
MultiAxisDevice is rotated, but collisions should cause rotations about
the mass center.