Fix Box Capsule Dcd Contact
The box capsule dcd contact was giving the incorrect penetration point
for the capsule, ie was adding the penetration when it should have been
subtracting it. This fix also behaves better as it finds one contact
between the box and capsule instead of two. This could result in
competing contacts that pull the objects in different directions to
resolve penetration, and in practice resulted in jumpy behavior.