LabJackDevice can now initialize timers. Initializing timers is necessary for enabling z-phase support for quadrature (i.e., reading encoders), as well as determining the outputs of PWM and frequency-output timers before the first update of the LabJackThread that is able to successfully pullOutput from the OutputProducer.
With z-phase support, the encoder provides a third line, the index,
that goes high to signal when the count should reset.

Note that the initial value must be sent at the same time as the mode
to enable z-phase support.
For example, for the high-level driver, trying consecutive calls
to ePut with TIMER_MODE and then TIMER_VALUE fails; instead we must
AddRequest and then GoOne.