Calculate the force at the VTCs attachment point
Calculate the force of the vtc at the attachment point and add the
contribution of the force at a distance to the torque. Before, the
application point was only being considered for the output force, this
makes the force and torque being sent back to the device is equal and
opposite of the force/torque being applied to the physics
representation.

Also, changed the attachment point to an optional parameter similar to
the stiffness and damping terms. So, if the attachment point is not
specified it is set to the center of mass of the physics representation
instead of 0,0,0.

Finally, the output force and torque staling is removed from the vtc.
These parameters are device dependant, and should be part of a device
filter.