Remove Asset::setFileName(), update Asset::load()
Asset::setFileName() barely duplicates the functionality of
Asset::load(), no need to keep it.

Update signature of Asset::load() to return nothing, it now asserts if
the given file name is not found and if the loading is failed.

Since std::bind() can't handle the overloading Asset::load() functions,
serialization of MeshShape and OctreeShape must be manually done.

Also update UnitTests because of above changes.