A VirtualToolCoupler with attachment point can disable inertial torques.
We need to handle the possibility that the input device's pose does not
correspond to the mass center, but some other point like the tip
of the tool.  The attachment point handles that possibility but also
adds torques that simulate inertia. Those torques are unwanted if the
input device has a replica of the tool for the handle so there is
already non-simulated inertia.