PhysicsManagerTests only starts runtime for InitTest.
Other tests should explicitly start the runtime and ensure an
update if they want to test the update code for some situation.

Otherwise (if the runtime is always running), a test may inadvertently
put the PhysicsManager in a situation that causes an assert
sporadically based on when the update occurs.