LabJackScaffold, catch an error condition earlier: bad timer numbers.
Timer numbers must be consecutive.  For example, it is invalid to
enable timer #1 if timer #0 is not enabled.  The result will be an
error in configureTimers, but the error message provided by the
device library is non-specific, e.g., on Windows: "Timer sharing error".

This commit adds a specific test for this issue, and logs the reason
that the configuration will fail.