Fix Physics Render Tests
The rigid body physics render tests were not working correctly. This was
caused by a change 13ab10ec2e5b31e8820194c78ac8e58a75e380f4, where the
rigid representation's state was set on doInitialize instead of
doWakeUp. If the pose of a an element being driven by a rigid
representation was set after doInitialize (as is done in these render
tests), it would be overwritten the first time the rigid representation
is updated. This change, put the setting of the rigid representation's
state back in doWakeUp so any changes to the scene element pose before
wakeup are accounted for.