Find yaml-cpp with ExternalProject_Add
Add a cmake variable called USE_SYSTEM_YAMLCPP. If this is ON, cmake
will try to find an installed version of yaml-cpp on the system. This is
how yaml-cpp has been found in the past. If USE_SYSTEM_YAMLCPP is OFF,
CMake will download yaml-cpp from our repository, configure, build, and
install yaml-cpp to the build directory of OpenSurgSim. This is the
default behavior, and makes it easier for new users, as they don't need
to find our forked version of yaml-cpp.