Use rigid addExternalGeneralizedForce API with localization in VTC
The external force added to the rigid use an application point, so the term torque += leverArm.cross(force) should not be included anymore.
But, to pass the information over to the haptic, this term needs to be present.

The rigid body is doing all the calculation already, but adding it to potentially already existing external forces, so we cannot just retrieve this information, it might not be what we need here.
Therefore, the calculation is redone (with optimization) locally in the VTC (torque update, stiffness/damping update) before being sent over to the haptic.