BoxCapsuleDcdContact uses brute force if the first search fails.
In the added test, doesIntersectBoxCapsule returns true, but
distancePointSegment returns a point that is outside the dilated box.

The closest point (on the capsule's segment) to the box center may not
be close enough to the box, even if the box and capsule are in contact
elsewhere.

This commit uses brute force in that case, testing all the box edges
against the capsule's segment.