Fix VirtualToolCoupler using ouput of RigidRep::addExternalGeneralizedForce
RigidRep::addExternalGeneralizedForce taking a localization has been modified to add extra terms to the force, stiffness and damping parameters.
This way, the caller can use this information on return. This is useful for the VTC for example.

the VTC code has been simplified, simply relying on the RigidRep::addExternalGeneralizedForce output.