Bugfix for BoxCapsuleDcdContact, setting capsule deepest point.
If the capsule's segment is completely outside the box, we calculate
segmentPoint as the closest point to the box's center.  If a line
segment starting at segmentPoint and aimed at the nearest point on the
box is less than the capsule radius, then we take that point on the box
as its deepest point, take that line as the normal, and calculate the
deepest capsule point based on the segmentPoint and normal.