Don't allow BUILD_SHARED_LIBS on Windows
Enabling shared libs on windows is going to require an EXPORT macro, and
exporting various stl containers (vector, map, etc). This isn't needed
with gcc, since everything is exported by default. So at the moment,
this change will display a cmake error if BUILD_SHARED_LIBS in ON, until
the needed changes are added.