Move SerializableStapling into ExampleStapling directory
Previously, ExampleStapling exported a library as well as an executable.
The library was named SurgSimExampleStapling (ugh), that was then used
by the seperate SerializableStapling project because it also needs the
StapleElement and StaplingBehavior. This change just moves the
SerialableStapling main file into ExampleStapling. And then the
CMakeLists.txt creates two executables, both including the shared
sources.