Moved Example Data directories to sub directories
We were inconsistently putting Data directories in the example
directory, and sometimes putting it in a common Examples/Data directory.
This moves all the Data directories to a sub-directory of each example.

Also, we now consistently copy the Data directory to the binary
directory. Before half the examples did this, and half didn't.