Make project name match directory name for examples
The project names for the examples were all pre-pended with Example, ie
ExampleInputVtc, etc. However, the directory names did not start with
Example (except ExampleStapling). Since they all live in the 'Examples'
directory and are located in the Examples folder in Visual Studio, I feel
this 'Example' is redundant, and removed it from the project names to
match the directory names.