Fix Octree and Constraint Generation w.r.t. local poses
The recent change to collision detection producing local positions
instead of global positions effected the way localizations are created
during contact constraint generation. Since the "local" position during
collision detection is local to the collision representation, it can't
be directly copied to the localization. The localization is "local" to
the physics representation. This change transforms the Location from the
collision representation's coordinate frame to the physics
representations coordinate frame before creating a localization in
ContactContstraintGeneration.

Also, using local positions broke the way the Octree creates contacts in
OctreeDcdContact. This change transforms the local contact position from
the individual nodes of the octree to the local position of the octree.