the yaml-cpp download now points to a tar.gz
For users that don't have git installed, pointing ExternalProject_Add to
a git repository will fail. It looks like cmake can extract tar.gz files
on its own, so we switch the link from a git repository to a download of
a tar.gz. The file is hosted on the OpenSurgSim Assembla page.