Tweak VTC parameters in Stapling example
Tweak the vtc parameters in the stapling example for better performance.
The VTC auto-calculates stiffness and damping parameters, but it assumes
the physics manager is running at 1000Hz. Since the stapling example
only runs at about 100Hz, these parameters need to be manually set for
better behavior.

Also, adjust the vtc parameter calculation to account for the attachment
point, when it calculates the moment of inertia.