Correct setInertialForces in VTC
The behaviour of the setInertialForces for the vtc is modified. Before
it incorrectly changed the calculation of the attacment point velocity.
Now, the vtc adds an external generalized force to the physics
representation at it's center of mass when calculateInertialForces is
off, and adds an external force at the attachment point when it is on.
This is the expected behaviour.