Set rate of PhysicsManager to 100Hz in Stapling example
Set the rate of the physics manager in the stapling example to 100Hz.
This is closer to the actual update rate. As a result, the vtc behaves
much better and the tissue behaves more realistically. This is because
before the physics was being run as if it was at 1kHz, but it wasn't. So
the mechanical properties were incorrect. As a result, the default vtc
parameters can be used again (although the defaults are tuned a bit
here).

Also, turn off the inertial force on the vtc in the stapling example.
This gives nicer tool behaviour.