Update keyboard toggling in Stapling example
Add the ability to turn on and off the visualize contacts behavour.  Set
the wireframes and the visualize contacts display to inactive by
default. Also improve the naming of some of the components to be more
consistent.

Removed the wire frame for the virtual teeth, I don't think it is
necessary.