Mass Center not itialized in RigidRepresentation
The mass center was not initialized in the Rigid Representation, as a
result, it was 0,0,0 on linux 64bit, and not on others. This caused
tests to incorrectly work on 64bit linux. When in fact, the center of
mass should be initialized to NaN like the other properties of the
rigid.  Then, when a shape is added, the mass center is calculated.

Also, added a shape to the rigid body used in some unit tests, so
that it would have a mass center, needed by the tests.