The LabJackDevice can now reset the hardware during destruct.
The hardware is designed to continue operation after communications
have closed, e.g., providing outputs.  This commit adds the ability
to reset the hardware, returning it to its default settings.
Note that resetting will cause USB re-enumeration and so will take a
few seconds before communications can be re-opened with the hardware.