Bugfix, pass the cmake_build_type to the externalproject for yaml.
Otherwise, when building with nmake jom, yaml will be build as Debug
even if the OSS build is set to Release, and thus cause a link
failure.

cmake_build_type is not used by MSVC, and cmake creates it for other
build systems, so we do not need to add it in SurgSimBuildFlags.