Adds DeviceInterface function to clear InputConsumers and OutputProducer
These function can be called to ensure the device object will not
attempt to pullOutput or pushInput to the rest of the scene.  Otherwise,
it is possible for a device to attempt to access a consumer or
producer that has already destructed.

Also, these functions are called in the destructor, which means the
mutex lock will be used to keep the removals threadsafe prior to actual
destruction.