Revise Accessible::decode() to handle ignored properties.
Add a 2nd parameter to Accessible::decode() as the list of ignored
properties.

The list is defaulted to empty.

Also, revised Accessible::decode() to warn when a property from the file
does not have a corresponding decoder. It also outputs a INFO when a
property with empty value is supplied.