Fix getChannel in the Image class
The getChennel was not working correctly due to the Eigen Map class.
When assigning values to it, ie:

  image->getChannel(0) = ...

The outer stride was incorrect, ie, it was just the number of rows (for
column major storage), when it should be the number of rows times the
inner stride (the number of channels in this case). I expected using the
InnerStride class would have calculated the outerstride on the fly. So
instead, the Eigen::Stride class is used here, and we calculated the
outer and inner stride.

This fix also contains unit tests to demonstrate the error.