Move the library header generation into own function
Move the cmake code to automatically create a library header into it's
own function (surgsim_add_library_header). Adding these headers probably
shouldn't be done for all libraries, such as the devices. So a seperate
cmake function is added here, and explicitly called in the libraries
as needed.

Also, modify the function to automatically figure out the
HEADER_DIRECTORY, and do a check to make sure the header you are trying
to create doesn't already exist in the library.