Fix Fem2DElementTriangle membrane mass matrix calculation
The membrane mass matrix is defined per axis (no coupling across axis), but was wrongly integrated in the code, creating coupling between the different axis.
Also the membrane only involves the axis X and Y. The axis Z contribution should come from the plate equation, not the membrane.
The reference to Przemieniecki book "Theory of Matrix Structural Analysis" says that the same mass matrix as axis X and Y can be used for the Z axis, but is not correct and can lead to wrong behaviors.