Update debug levels and message in OpenNI device
Change some debugging levels to SEVERE, since they occur when the OpenNI
camera is unusable. Added some debug info to a couple of asserts.

Also, removed a check of how many devices are connected. This was added
for debug purposes and isn't needed.

Added doxygen comment for the return value os
OpenNIDevice::isInitialized