Fix doxygen issue with static_assert
It appears that doxygen 1.8.8 has a problem with static_assert. ie, it
can't find declarations for functions defined after static_assert. To
sidestep this issue, this changes the template parameter (BlockSize)
from int to size_t. So there is no longer a need to call static_assert
to make sure BlockSize is greater than zero.