Set default build back to Release builds
It used to be that the default build in non-multiconfiguration
generators (ie, not Visual Studio) was set explicitly to "Release" in
SurgSimBuildFlags. This was removed, because it was thought to be not
neccassary in 7ffda85ee162. But in fact, the default build type in CMake
is Debug. This affected single configuration generators (Make,
NMake, Jom, Ninja, etc).

This change re-adds the setting of CMAKE_BUILD_TYPE to Release if it is
not set. But also checks that it is not a multiconfiguration generator
(NOT CMAKE_CONFIGURATION_TYPES), where this setting is unneccassary (and
conflicts with the way we currently pass the build type in
External_yamlcpp.cmake).