Fix fem2d plate mass matrix calculation
The plate did not take into account the main contribution, which is the contribution of the dof z (displacement w).
Plus the contribution of the dofs thetax and thetay was wrong, The dot product between the vector of functions was inverted producing a single 2x2 matrix instead of 2 9x9 matrices.

The 3 contributions have been split in 3 methods to keep the code clear and clean.