Fix link to MathJax and add a FindMathJax.cmake
Fix a broken link the the MathJax javascript library, needed when
building the html documentation. Also, add a FindMathJax.cmake and use
find_package to find a local MathJax installation. It will look in
standard locations and in the environment/cmake variable MATHJAX_DIR.

Update README to list MathJax as an option dependency.