Have doxygen give location of non-local MathJax
Kind of a sneeky way to remove the MATHJAX_RELPATH option from the
Doxyfile when there is no local copy. When this parameter is missing,
doxygen provides the link to the online version of MathJax.js. This
leaves us out of the business of providing this link, and keeping it up
to date.