NovintScaffold follows the canonical sequence for HDAL function calls.
1) hdlInit* for each device
2) hdlStart
3) hdlCreateServoOp
...
4) hdlDestroyServoOp
5) hdlStop
6) hdlUninitDevice for each device

This sequence is provided in the Programmer's Guide.
With the previous approach, hdl calls would fail or exit the
application in the case of multiple NovintDevices.

Since the NovintScaffold does not know which devices will be needed
at the time that the first device is registered, and we do not want to
prevent registering additional devices at any time, we need to
hdlInit* every connected device before calling hdlStart.
Then, when registerDevice is called it simply matches the device
handles with the name.  That matching is done via a configuration file.
Finally, when the Scaffold destructs we deleteServoOp, stop, & uninit.

Using an empty string for the initialization name will register to
the first available (initialized) device.  If a non-empty string is
used, findHandle will use the std::map<string, string> loaded from
novint.ini (a local file holding a yaml node) to get the serial number
that should be used for that name.

Note, this commit fails the NovintScaffoldTests...
probably because the scaffold is holding objects with shared pointers
to itself.