NovintScaffold is a singleton.
To ensure that we follow the canonical sequence for calling HDAL
functions (e.g., only calling hdlStart once after all the calls to
hdlInit*), there can be only one NovintScaffold object throughout
the lifetime of the application.  A SharedInstance is not
appropriate because it destructs when no more devices exist, and
a new one is constructed when a new device is created.