Removes NovintScaffold::setDefaultLogLevel.
The log level should be set using the standard logger functions,
since the loggers are accessible from anywhere, not owned by the
scaffold object.

Also, the NovintScaffold now always uses the same logger name instead
of accepting a logger in its constructor (which was never used), and
the logger name is "Devices/Novint" so that
LoggerManager::setThreshold("Devices/", someThreshold) will
get the NovintScaffold logger (and hopefully eventually all devices'
loggers).