Revise LabJack unit test to expose a bug in LabJackScaffold
No matter there is an output producer set on a LabJack device or not,
the LabJack device should work.

With this revision, it exposed a bug in LabJackScaffold that when an
output producer is set on a LabJack device but not providing any initial
output, input consumers of this LabJack won't get any input.