Add a Pose to Input and Output Components for alignment
Make InputComponents and OutputComponents Representations, so they now
have a pose. This allows input and output to be "aligned" or
"registered" to the scene from within the simulation. Device specific
alignment should still be done with the PoseTransform device filter.

With this change, behaviors that use the input/output components need to
also use it's pose, if the input/output data is "pose"-able. This change
adds these calculations to the DriveElementFromInputBehavior and the
VirtualToolCoupler, the only current behaviors that deal with pose data
from devices. These two behaviors now transform the input/output data
with this transform.

Now, to align an input, you just set a transform on the InputComponent
(with setLocalPose). To align an output, you set a transform on the
OutputComponent (with setLocalPose).