Handle conversion of containers in the BufferedValue
Create two version of a private doPublish, that is selected based on if
the contained type is a container or not. If it is container, it does a
container specific copy. To allow for T to be copied to ST if they are
different container types, but their contained type is compatible.  This
is used to copy a container of ParticleReference to a container of
Particle in the Particles library.