Commit the responses to the merge request
This commit takes care of most of the code changes to the base files with
the exception of renaming this class SamplingMetrics and providing testing
code. These additional changes will follow.

The MetricBase class was renamed to SamplingMetricBase to allow for the
possibility of a superclass with additional metric functionality such as
reasoning on a series of measurements or of making broader go/nogo
decisions.

Added in unit tests to test the SamplingMetricBase class. The tests use an
additional mock class defined in MockObjects.h and required some (minor)
refactoring of the SamplingMetricBase class.