Adds a KalmanFilter class, for linear estimation without control inputs.
The VelocityFromPose filter is working, though the values have not
been adjusted for optimal performance.

For the angular velocity filter, it's calculating the delta orientation,
and dividing by the period to get an angular velocity, then passing it
to a KalmanFilter on the velocity and acceleration (a constant
acceleration model.

This works, but it would probably be better to
directly use the orientation, e.g., measuring quaternions and using a
state of {quaternion, angular velocity, angular acceleration}, but using
orientations is tricky because they are SO(3)
instead of R3.  The majority of implementations use
Extended Kalman Filters (or more advanced variations), and no
example explains how to handle when the measurement jumps between
states, i.e., Quaterniond(pose.linear()).vec()
goes from (0,0,0.99) -> (0,0,-0.99).